+++ react/src/js/components/Consonant/Container/Container.jsx
@@ -71,6 +71,7 @@ import {
     transformFiltersWithCategories,
     expandGroupFiltersToChildren,
     getGroupedFilterSelections,
+    isMobile,
 } from '../Helpers/Helpers';
 
 
@@ -1756,7 +1757,7 @@ const Container = (props) => {
     }, []);
 
     const carouselClass = classNames({
-        'modern-carousel': isModernCarousel,
+        'modern-carousel': isModernCarousel || (isCarouselContainer && isMobile()),
         'modern-carousel--light': isLightCarousel,
     });
 
@@ -1934,7 +1935,7 @@ const Container = (props) => {
                                 resQty={gridCardLen}
                                 cards={gridCards}
                                 cardStyle={cardStyle}
-                                carouselType={isModernCarousel ? 'modern' : 'default'}
+                                carouselType={(isModernCarousel || isMobile()) ? 'modern' : 'default'}
                                 role="tablist"
                                 onCardBookmark={handleCardBookmarking} />
                             }